+++ fastcrypto/src/lib.rs
@@ -22,8 +22,8 @@ pub mod signature_tests;
 pub mod ed25519_tests;
 
 #[cfg(test)]
-#[path = "tests/secp256k1_tests.rs"]
-pub mod secp256k1_tests;
+#[path = "tests/secp256k1_recoverable_tests.rs"]
+pub mod secp256k1_recoverable_tests;
 
 #[cfg(test)]
 #[path = "tests/secp256r1_tests.rs"]
